% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathview.2.R
\name{pathview.2}
\alias{pathview.2}
\title{this is a function to run pathview.2 function}
\usage{
pathview.2(
  run,
  diff.tool,
  gene.data = cnts,
  ref,
  samp,
  gsets,
  pathway.id,
  both.dirs = list(gene = T, cpd = T)
)
}
\description{
this is a function to run pathview.2 function
}
