% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_path_analysis.R
\name{run_path_analysis}
\alias{run_path_analysis}
\title{This is the function to run pathway analysis}
\usage{
run_path_analysis(lf.c, gsets)
}
\description{
This is the function to run pathway analysis
}
