% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundifftool.R
\name{rundifftool}
\alias{rundifftool}
\title{this is a function to run different diff expression analysis tools}
\usage{
rundifftool(diff.tool, gene.data, ref, samp, outname)
}
\description{
this is a function to run different diff expression analysis tools
}
